fixed breaking due to missing directory on sacc_generator
